requirements.txt
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
eventlet==0.33.3
Flask==3.0.0
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.0
WTForms==3.1.0
